fix: working /var /etc /home
This provides a var.mount and rootfs symlinks to /var/DIRS.

It is a bit of a hacky solution and there are probably more
elegant ways to do this with mkosi, but a few things stick out:

1) sometimes /var was getting mounted magically due to a mount
unit generated by systemd due to discoverable partitions spec,
but sometimes it was not... per the spec, installers are still
recommended to create fstab/mount units, so it seems appropriate
to add var.mount

2) sharing the same mkosi trees for both live image and target
installed partitions can create issues... for example, i had to
place the var.mount in a tree not used by the live image else
it would never boot because the expected PARTLABEL was not found.

3) i staged symlinks in rootfs (with the nonlivefs dir) in a
pattern which matches bootc's approach. even /etc is a symlink
to /var/etc here.